Generate augmentation data with on disk storage

Generate augmentation data with on disk storage

generateDiskData(
  inputImageList,
  segmentationImageList,
  segmentationNumbers,
  selector,
  addCoordConv = 0,
  segmentationsArePoints = FALSE,
  maskIndex,
  numpynames,
  numberOfSimulations = 16,
  referenceImage = NULL,
  transformType = "rigid",
  noiseModel = "additivegaussian",
  noiseParameters = c(0, 0.002),
  sdSimulatedBiasField = 5e-04,
  sdHistogramWarping = 5e-04,
  sdAffine = 0.2
)

Arguments

inputImageList
list of lists of input images to warp. The internal list sets contains one or more images (per subject) which are assumed to be mutually aligned. The outer list contains multiple subject lists which are randomly sampled to produce output image list.

segmentationImageList
of segmentation images corresponding to the input image list.

segmentationNumbers
the integer list of values in the segmentation to model

selector
subsets the inputImageList and segmentationImageList (eg to define train test splits)

addCoordConv
boolean - generates another array with CoordConv data

segmentationsArePoints
boolean - converts segmentations to points

maskIndex
the entry within the list of lists that contains a mask

numpynames
the names of the numpy on disk files should contain something with the string mask if using maskIndex and something with the word coordconv if using CC

numberOfSimulations
number of output images. Default = 10.

referenceImage
defines the spatial domain for all output images. If the input images do not match the spatial domain of the reference image, we internally resample the target to the reference image. This could have unexpected consequences. Resampling to the reference domain is performed by testing using antsImagePhysicalSpaceConsistency then calling resampleImageToTarget upon failure.

transformType
one of the following options c( "translation", "rigid", "scaleShear", "affine"," deformation" , "affineAndDeformation" ).

noiseModel
one of the following options c( "additivegaussian", "saltandpepper", "shot", "speckle" )

noiseParameters
'additivegaussian': c( mean, standardDeviation ), 'saltandpepper': c( probability, saltValue, pepperValue) , 'shot': scale, 'speckle': standardDeviation. Note that the standard deviation, scale, and probability values are max values and are randomly selected in the range [0, noise_parameter]. Also, the "mean", "saltValue" and pepperValue" are assumed to be in the intensity normalized range of [0, 1].

sdSimulatedBiasField
Characterize the standard deviation of the amplitude.

sdHistogramWarping
Determines the strength of the bias field.

sdAffine
Determines the amount of transformation based change

Value

list of array

Author

Avants BB

Examples

mydf = NULL